refactor: 도메인 로직으로 분리
#130
